Move the UnpackerMessages and tests to a separate object
This is just a string builder, and doesn't need to be mixed in with I/O.
