+++ bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/services/Unpacker.scala
@@ -1,7 +1,6 @@
 package uk.ac.wellcome.platform.archive.bagunpacker.services
 
 import java.io.InputStream
-import java.text.NumberFormat
 import java.time.Instant
 
 import grizzled.slf4j.Logging
@@ -37,12 +36,6 @@ trait Unpacker extends Logging {
 
   def formatLocation(location: ObjectLocation): String
 
-  def createMessage(summary: UnpackSummary): String = {
-    val displayFileCount = NumberFormat.getInstance().format(summary.fileCount)
-    s"Unpacked ${summary.size} from $displayFileCount file${if (summary.fileCount != 1) "s"
-    else ""}"
-  }
-
   def unpack(
     ingestId: IngestID,
     srcLocation: ObjectLocation,
@@ -64,7 +57,7 @@ trait Unpacker extends Logging {
         Success(
           IngestStepSucceeded(
             summary,
-            maybeUserFacingMessage = Some(createMessage(summary))
+            maybeUserFacingMessage = Some(UnpackerMessage.create(summary))
           )
         )
